Setup Firebase features boilerplate
This cl sets up the boilerplate for a bunch of Firebase features.

I checked off a bunch of the features that I believe I am going to use:

Firestore
Functions
Hosting
Emultators
Storage

A lot of this will likely not work correctly with nextjs, but I'll resolve
that much later.