Create all the components necessary for the game
This cl creates all the components needed for the Wordle game. Whats not
included is implementing the game.

The overarching plan is to persist the state of a game session in firebase.

In order to do that, I need to wire in Firebase Auth and Firebase datastore.

... Still playing around with those two. Will follow up once I've nailed down
a design that is flexible enough.