Connect Firestore to ui, rendering an in-progress wordle game
This cl renders an in-progress wordle game. The data model consists of
a game documents composed of a list of repeated attempts, a list of the
hints provided for each of those attempts, and the information needed for
the current attempt.

Tomorrow I'll work on interactions and getting the game to actually work.

Stay tuned.