Convert to ts (as a poor mans replacement for an ORM)
This cl converts all the function methods to ts as a cheap replacement ORM.

One "peeve" of mine, is that I don't really know the typing of the data I
am dealing with. I know, its a no-sql database, etc...

Using typescript gives the slightest bit of semblence of trying to get
the typing to work as I would expect.

It doesn't do a perfect job, but thats ok, its easier this way.